extract EvalError into a separate module;
