+++ ergotree-interpreter/src/eval.rs
@@ -1,28 +1,18 @@
 //! Interpreter
-use bounded_vec::BoundedVecOutOfBounds;
 use ergotree_ir::mir::constant::TryExtractInto;
 use ergotree_ir::pretty_printer::PosTrackingWriter;
 use ergotree_ir::pretty_printer::Print;
 use ergotree_ir::sigma_protocol::sigma_boolean::SigmaProp;
-use ergotree_ir::source_span::Span;
-use sigma_ser::ScorexParsingError;
-use sigma_ser::ScorexSerializationError;
 use std::rc::Rc;
 
-use ergotree_ir::ergo_tree::ErgoTreeError;
-use ergotree_ir::mir::constant::TryExtractFromError;
 use ergotree_ir::mir::expr::Expr;
 use ergotree_ir::mir::value::Value;
-use ergotree_ir::serialization::SigmaParsingError;
-use ergotree_ir::serialization::SigmaSerializationError;
 use ergotree_ir::sigma_protocol::sigma_boolean::SigmaBoolean;
 
 use cost_accum::CostAccumulator;
 use ergotree_ir::types::smethod::SMethod;
-use thiserror::Error;
 
 use self::context::Context;
-use self::cost_accum::CostError;
 use self::env::Env;
 
 /// Context(blockchain) for the interpreter
@@ -60,6 +50,7 @@ pub(crate) mod decode_point;
 mod deserialize_context;
 mod deserialize_register;
 pub(crate) mod downcast;
+mod error;
 pub(crate) mod exponentiate;
 pub(crate) mod expr;
 pub(crate) mod extract_amount;
@@ -104,87 +95,7 @@ pub(crate) mod val_use;
 pub(crate) mod xor;
 pub(crate) mod xor_of;
 
-/// Interpreter errors
-#[derive(Error, PartialEq, Eq, Debug, Clone)]
-pub enum EvalError {
-    /// AVL tree errors
-    #[error("AvlTree: {0}")]
-    AvlTree(String),
-    /// Only boolean or SigmaBoolean is a valid result expr type
-    #[error("Only boolean or SigmaBoolean is a valid result expr type")]
-    InvalidResultType,
-    /// Unexpected Expr encountered during the evaluation
-    #[error("Unexpected Expr: {0}")]
-    UnexpectedExpr(String),
-    /// Error on cost calculation
-    #[error("Error on cost calculation: {0:?}")]
-    CostError(#[from] CostError),
-    /// Unexpected value type
-    #[error("Unexpected value type: {0:?}")]
-    TryExtractFrom(#[from] TryExtractFromError),
-    /// Not found (missing value, argument, etc.)
-    #[error("Not found: {0}")]
-    NotFound(String),
-    /// Register id out of bounds
-    #[error("{0}")]
-    RegisterIdOutOfBounds(String),
-    /// Unexpected value
-    #[error("Unexpected value: {0}")]
-    UnexpectedValue(String),
-    /// Arithmetic exception error
-    #[error("Arithmetic exception: {0}")]
-    ArithmeticException(String),
-    /// Misc error
-    #[error("error: {0}")]
-    Misc(String),
-    /// Sigma serialization error
-    #[error("Serialization error: {0}")]
-    SigmaSerializationError(#[from] SigmaSerializationError),
-    /// Sigma serialization parsing error
-    #[error("Serialization parsing error: {0}")]
-    SigmaParsingError(#[from] SigmaParsingError),
-    /// ErgoTree error
-    #[error("ErgoTree error: {0}")]
-    ErgoTreeError(#[from] ErgoTreeError),
-    /// Not yet implemented
-    #[error("evaluation is not yet implemented: {0}")]
-    NotImplementedYet(&'static str),
-    /// Invalid item quantity for BoundedVec
-    #[error("Invalid item quantity for BoundedVec: {0}")]
-    BoundedVecError(#[from] BoundedVecOutOfBounds),
-    /// Scorex serialization error
-    #[error("Serialization error: {0}")]
-    ScorexSerializationError(#[from] ScorexSerializationError),
-    /// Scorex serialization parsing error
-    #[error("Serialization parsing error: {0}")]
-    ScorexParsingError(#[from] ScorexParsingError),
-    /// Wrapped eval error with environment after evaluation
-    #[error("eval error: {error}, env: {env:?}")]
-    WrappedWithEnvError {
-        /// eval error
-        error: Box<EvalError>,
-        /// environment after evaluation
-        env: Env,
-    },
-    /// Wrapped eval error with source span
-    #[error("eval error: {error}, source span: {source_span:?}")]
-    WrappedWithSpan {
-        /// eval error
-        error: Box<EvalError>,
-        /// source span
-        source_span: Span,
-    },
-}
-
-impl EvalError {
-    /// Wrap eval error with source span
-    pub fn wrap_with_span(self, source_span: Span) -> Self {
-        EvalError::WrappedWithSpan {
-            error: Box::new(self),
-            source_span,
-        }
-    }
-}
+pub use error::EvalError;
 
 /// Result of expression reduction procedure (see `reduce_to_crypto`).
 #[derive(PartialEq, Eq, Debug, Clone)]
